requirements.txt
pyqrcode
pyTelegramBotAPI
colorama
Pillow
pypng
zbar
pyzbar
